Now we're thinking with Fibers.
The multiprocess model was a little overkill, and impacted performance for my
use case, here we've switched it out for Fibers.

As we can't return from the Net::HTTP.start block, or even the request block, we
must find a way of continuing without returning. Fibers allow us to effectively
jump back to the http block, and get the next value out of it.